Chore: Displaying correct profile data when clicking on both logged in user and other users
