+++ account/index.html
@@ -67,7 +67,7 @@
                     </div>
                     <div class="flex flex-col bg-ui-white items-center gap-5 w-[350px] h-[168px] border rounded-sm p-4 justify-center">
                         <img class="w-10 h-10" src="../public/assets/icons/account/open-profile.svg" alt="open profile icon">
-                        <button class="w-[172px] h-[38px] font-text bg-primary-green text-ui-white border border-transparent rounded-sm text-sm hover:bg-ui-white hover:border-primary-green hover:text-primary-green">
+                        <button id="open-profile-page" class="w-[172px] h-[38px] font-text bg-primary-green text-ui-white border border-transparent rounded-sm text-sm hover:bg-ui-white hover:border-primary-green hover:text-primary-green">
                             Open Profile Page
                         </button>
                     </div>
